% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds_spp}
\alias{birds_spp}
\title{Birds species list}
\format{
\subsection{\code{birds_spp}}{

A \code{vector} with 9993 bird species.
}
}
\source{
\if{html}{\out{
<??>
}}
}
\usage{
birds_spp
}
\description{
description
}
\keyword{datasets}
