% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRangeThrough.R
\name{calcRangeThrough}
\alias{calcRangeThrough}
\title{Calculate paleo diversity curves using the Range Through method}
\usage{
calcRangeThrough(data, tax_lvl = "species")
}
\arguments{
\item{data}{A \code{data.frame} containing the columns: \code{max_ma}, \code{min_ma}
and  the name provided in \code{tax_lvl}. \code{max_ma} and \code{min_ma} are
respectively the early and late bounds of rock layer's age. \code{tax_lvl} column
is the taxonomic level of the data. Any additional columns are ignored.}

\item{tax_lvl}{A \code{character} giving the taxonomic in which calculations
will be based on (default value is \code{"species"}). This must refer to the
column names in \code{data}.}
}
\value{
A \code{data.frame} containing the diversity (column \code{div}) of
the chosen taxonomic level through time. Time moments are the
layer boundaries given in \code{data}.
}
\description{
\code{calcRangeThrough} calculates richness through time using the method of
Range Through (ADD REFERENCE).
}
\examples{

data("trilob_fossil")
div <- calcRangeThrough(trilob_fossil)
plot(x=div$age, y=div$div, type="l", 
     xlab = "Time (Mya)", ylab = "Richness", 
     xlim=rev(range(div$age)))#' 

}
\references{
ADD REFERENCE
}
\author{
Matheus Januario, Jennifer Auler
}
