% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStdMethod.R
\name{calcStdMethod}
\alias{calcStdMethod}
\title{Calculate paleo diversity curves using the standard method}
\usage{
calcStdMethod(data, tax_lv = "species", bin_reso = 1)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns: \code{max_ma}, \code{min_ma}
and  the name provided in \code{tax_lvl}. \code{max_ma} and \code{min_ma} are
respectively the early and late bounds of rock layer's age. \code{tax_lvl} column
is the taxonomic level of the data. Any additional columns are ignored.}

\item{tax_lv}{A \code{character} giving the taxonomic in which calculations
will be based on (default value is \code{"species"}). This must refer to the
column names in \code{data}.}

\item{bin_reso}{A \code{numeric} assigning the resolution (length) of the
time bin to consider in calculations. Default value is \code{1} (which in
most cases - e.g. those following the paleoBiology Database default
timescale - will equate to one million years)}
}
\value{
A \code{data.frame} containing the diversity (column \code{div}) of
the chosen taxonomic level, through time - with time moments being a
sequence of arbitrary numbers based on \code{bin_reso}
}
\description{
\code{calcRangeThrough} calculates richness through time using the
standard method (ADD REFERENCE).
}
\examples{

data("trilob_fossil")
div <- calcStdMethod(trilob_fossil)
plot(x=div$age, y=div$div, type="l", 
     xlab = "Time (Mya)", ylab = "Richness", 
     xlim=rev(range(div$age)))

# Now same plot, but using a different bin resolution:
data("trilob_fossil")
div2 <- calcStdMethod(trilob_fossil, bin_reso = 3)
plot(x=div2$age, y=div2$div, type="l", 
     xlab = "Time (Mya)", ylab = "Richness", 
     xlim=rev(range(div$age)))

}
\references{
ADD REFERENCE
}
\author{
Matheus Januario, Jennifer Auler
}
