% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytOxidase}
\alias{cytOxidase}
\title{Cytocrome Oxidase sequences}
\format{
\subsection{\code{cytOxidase}}{

A \code{vector} with 17 entrys
\describe{
\item{names}{Organism}
\item{sequence}{Aminoacid sequence}
}
}
}
\source{
<??>
}
\usage{
cytOxidase
}
\description{
description
}
\keyword{datasets}
