% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dinos_fossil}
\alias{dinos_fossil}
\title{Occurrence of dinossaur fossils}
\format{
\subsection{\code{dinos_fossil}}{

A \code{data.frame} with 13 colums and 15527 rows.
\describe{
\item{phylum}{phylum}
\item{class}{class}
\item{order}{order}
\item{family}{family}
\item{genus}{genus}
\item{species}{species}
\item{early_interval}{earlier known geological period of ocurence}
\item{late_interval}{later known geological period of ocurence}
\item{max_ma}{ocurrence's oldest time boundary in million years}
\item{min_ma}{ocurrence's newest time boundary in million years}
\item{midpoint}{midpoint between max_ma and min_ma}
\item{lng}{longitude of place where occurrence was found}
\item{lat}{latitude of place where occurrence was found}
}
}
}
\source{The Paleobiology Database (donwloaded on 2022-03-11)

Data URL: "http://paleobiodb.org/data1.2/occs/list.csv?datainfo&rowcount&base_name=Dinosauria&show=full,classext,genus,subgenus,acconly,ident,img,etbasis,strat,lith,env,timebins,timecompare,resgroup,ref,ent,entname,crmod"
\if{html}{\out{
<??>
}}
}
\usage{
dinos_fossil
}
\description{
A data frame contaning many dinosaur (including avian species) fossil occurrences from different moments of the geological past. Much information (i.e., extra columns) was removed from the original datset to make it more compact, but it can be accessed by the data URL.
}
\keyword{datasets}
