% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lttPlot.R
\name{lttPlot}
\alias{lttPlot}
\title{Make a lineage trough time (LTT) plot}
\usage{
lttPlot(phy, lwd = 1, col = "red", PLOT = T, rel_time = F, add = F)
}
\arguments{
\item{phy}{A \code{phylo} object, as specific by the \code{ape} package.}

\item{lwd}{Line width.}

\item{col}{Line color.}

\item{PLOT}{A \code{logical} indicating with calculations should be plotted.
If \code{FALSE}, function returns a list of the calculated points.}

\item{rel_time}{A \code{logical} indicating if time should be calculated in
absolute scale. If \code{FALSE} (default), plots relative time since \code{phy}'s
crown age.}

\item{add}{A \code{logical} indicating if plot should be added to
pre-existing plot. Default is \code{FALSE}.}
}
\value{
If \code{PLOT = FALSE}, a list the richness of each point in time,
and \code{phy}'s crown age.
}
\description{
\code{lttPlot} plots the lineage trough time (LTT) of a \code{phylo} object.
It also adds a reference line connecting the edges of the graph.
}
\examples{

S <- 1
E <- 0
set.seed(1)
phy <- simulateTree(pars = c(S, E), max.taxa = 20)
lttPlot(phy)
lttPlot(phy, PLOT = FALSE)

}
\references{
Paradis, E. (2012). Analysis of Phylogenetics and Evolution with R (Vol. 2).
New York: Springer.
}
\author{
Daniel Rabosky, Matheus Januario, Jennifer Auler
}
