% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatSelSim.R
\name{NatSelSim}
\alias{NatSelSim}
\title{Simulating natural selection through time in a bi-allelic gene}
\usage{
NatSelSim(
  w11 = 1,
  w12 = 1,
  w22 = 0.9,
  p0 = 0.5,
  NGen = 10,
  plot_type = "animateall",
  printData = FALSE
)
}
\arguments{
\item{w11}{Number giving the fitness of genotype A1A1. Values will be
normalized if any genotype fitness exceeds one.}

\item{w12}{Number giving the fitness of genotype A1A2. Values will be
normalized if any genotype fitness exceeds one.}

\item{w22}{Number giving the fitness of genotype A2A2. Values will be
normalized if any genotype fitness exceeds one.}

\item{p0}{Initial (time = 0) allelic frequency of A1.
A2's initial allelic frequency is \code{1-p0}.}

\item{NGen}{Number of generation that will be simulated.}

\item{plot_type}{String indicating if plot should be animated.
The default, "animateall", animate all possible panels.
Other options are "static", "animate1", "animate3", or "animate4".}

\item{printData}{Logical indicating whether all simulation results should be
returned as a \code{data.frame}. Default value is \code{FALSE}.}
}
\value{
A \code{data.frame} containing the number of individuals for each
genotype.
}
\description{
\code{NatSelSim} simulates natural selection in a bi-allelic gene through
\code{NGen} generations.
}
\details{
If any value of fitness (i.e., \code{w11}, \code{w12},
\code{w22}) is larger than one, fitness is interpreted as absolute fitness
and values are re-normalized.
}
\examples{

#using the default values (w11=1, w12=1, w22=0.9, p0=0.5, NGen=10)
\dontrun{NatSelSim()}

# Continuing a simulation for extra time:
# Run the first simulation
sim1=NatSelSim(w11 = .4, w12 = .5, w22 = .4, p0 = 0.35, NGen = 5, plot_type = "static", printData = TRUE)

# Then take the allelic frequency form the first sim:
new_p0 <- (sim1$AA[nrow(sim1)] + sim1$Aa[nrow(sim1)]*1/2) 
# and use as p0 for a second one:

NatSelSim(w11 = .4, w12 = .5, w22 = .4, p0 = new_p0, NGen = 5, plot_type = "static")


}
\author{
Matheus Januario, Jennifer Auler, Dan Rabosky
}
