% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{plotNatSel}
\alias{plotNatSel}
\title{Plot NatSelSim output}
\usage{
plotNatSel(
  gen_HW = gen_HW,
  p_t = p_t,
  w_t = w_t,
  t = t,
  W_gntp = c(w11, w12, w22),
  plot_type = "animateall"
)
}
\arguments{
\item{gen_HW}{Dataframe with A1A1, A1A2 and A2A2 genotypic
frequencies in each generation (nrows = NGen)}

\item{p_t}{Allelic frequency through time}

\item{w_t}{Mean population fitness through time}

\item{t}{time}

\item{W_gntp}{Initial genotypic fitness}

\item{plot_type}{String indicating if plot should be animated.
The default, "animateall", animate all possible panels.
Other options are "static", "animate1", "animate3", or "animate4".}
}
\value{
Plot of NatSelSim's output (see \code{NatSelSim}'s help page for
details).
}
\description{
Plot NatSelSim output
}
