% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{plotWFDrift}
\alias{plotWFDrift}
\title{Plot WFDriftSim output}
\usage{
plotWFDrift(p_through_time, plot_type = plot, knitr = FALSE)
}
\arguments{
\item{p_through_time}{Matrix with nGen columns and nSim lines}

\item{plot_type}{String. Options are "static" or "animate"}

\item{knitr}{Logical indicating if plot is intended to show up in RMarkdown files made by the \code{Knitr} R package.}
}
\value{
A static or animated plot of populations under genetic drift through time
}
\description{
Plot WFDriftSim output
}
\examples{
store_p = WFDriftSim(Ne = 5, nGen = 10, p0=.2, nSim=5, plot = "none", printData = TRUE)
plotWFDrift(store_p, "static")
}
