% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whale_body_size}
\alias{whale_body_size}
\title{Whale body size}
\format{
\subsection{\code{whale_body_size}}{

A \code{data.frame} with 4 colums and 75 rows.
\describe{
\item{species}{species}
\item{log_mass}{log of body mass (?units)}
\item{S}{?}
\item{color}{?}
}
}
}
\source{
<??>
}
\usage{
whale_body_size
}
\description{
description
}
\keyword{datasets}
