% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whale_phylo}
\alias{whale_phylo}
\title{Whale Phylogeny}
\format{
An ultrametric \code{phylo} object with 87 tips
}
\source{
Original phylogeny generation by Steeman et al (2001).
File obtained from Rabosky et al, 2014.
}
\usage{
whale_phylo
}
\description{
An ultrametric phylogenetic tree of the living cetaceans. Phylogeny
generated by Steeman et al (2009). This dataset is part of the package and is licensed
under the Creative Commons Attribution 4.0 International License (CC BY 4.0).
}
\references{
Rabosky, D. L., Grundler, M., Anderson, C., Title, P., Shi, J. J., Brown, J. W., ... & Larson, J. G. (2014). BAMM tools: an R package for the analysis of evolutionary dynamics on phylogenetic trees. Methods in Ecology and Evolution, 5(7), 701-707.

Steeman, M. E., Hebsgaard, M. B., Fordyce, R. E., Ho, S. Y., Rabosky, D. L., Nielsen, R., ... & Willerslev, E. (2009). Radiation of extant cetaceans driven by restructuring of the oceans. Systematic biology, 58(6), 573-585.
}
\keyword{datasets}
