% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds_spp}
\alias{birds_spp}
\title{Birds species list}
\format{
A \code{vector} containing the names of almost all (i.e., 9993) extant bird species, following Jetz et al. (2012) taxonomy. This dataset is part of the package and is
licensed under the Creative Commons Attribution 4.0 International License (CC BY 4.0).
}
\source{
Actual file downloaded from \url{https://vertlife.org/data/}
}
\usage{
birds_spp
}
\description{
Birds species list
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444-448.
}
\keyword{datasets}
