% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytOxidase}
\alias{cytOxidase}
\title{Cytochrome Oxidase sequences}
\format{
A  object of class "ProteinSeq" with 17 entries, each representing
a different animal species
\describe{
\item{names}{Organism's popular name or a taxonomic classification}
\item{sequence}{Aminoacid sequence of length 513}
}
}
\source{
Amino acid sequences were originally downloaded from genebank and later
curated and aligned by Daniel L. Rabosky.
}
\usage{
cytOxidase
}
\description{
\code{cytOxidase} is a set of homologous protein sequences from the GENE
cytochrome oxidase SUBUNIT 1 gene. This mitochondrial gene, often known as
CO1 (“see-oh-one”),  plays a key role in cellular respiration. C01 contains
approximately 513 aminoacids and has been used by previous studies for
reconstructing phylogenetic trees and estimating divergence times in
Metazoaria by assuming a molecular clock. Its 5' partition is used for the
‘Barcoding of Life’ initiative, for instance. This dataset is part of the
package and is licensed under the Creative Commons Attribution 4.0
International License (CC BY 4.0).
}
\details{
The object of class "ProteinSeq" is structured as a named
\code{vector} of 17 different animal species, with each individual component
being a sequence of 513 aminoacids.
}
\keyword{datasets}
