% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSeq.R
\name{is.proteinSeq}
\alias{is.proteinSeq}
\title{Details, generics, and methods for the \code{proteinSeq} class}
\usage{
is.proteinSeq(x)
}
\arguments{
\item{x}{an object of the class \code{proteinSeq}}
}
\description{
The \code{proteinSeq} class is an input for the functions \code{countSeqDiffs} and \code{is.proteinSeq}. It consists of a character vector. Each entry in this vector represents the amino acid (the protein components coded by a gene) sequence, for a given alligned protein sequence. The object must be a character, named vector, with the names tipically corresponding to the species (name could be scientific or common name) from which every sequence came. The characters within the vector must correspond to valid amino acid symbols (i.e. capitalized letters or deltion "_" symbols).
}
\examples{

data("cytOxidase")
is.proteinSeq(cytOxidase)

summary.proteinSeq(cytOxidase)

head.proteinSeq(cytOxidase)

tail.proteinSeq(cytOxidase)

}
\author{
Daniel Rabosky, Matheus Januario, Jennifer Auler
}
