% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSeq.R
\name{is.ProteinSeq}
\alias{is.ProteinSeq}
\alias{print.ProteinSeq}
\alias{summary.ProteinSeq}
\alias{head.ProteinSeq}
\alias{tail.ProteinSeq}
\title{Details, generics, and methods for the \code{proteinSeq} class}
\usage{
is.ProteinSeq(x)

\method{print}{ProteinSeq}(x)

\method{summary}{ProteinSeq}(x)

\method{head}{ProteinSeq}(x, n = 20)

\method{tail}{ProteinSeq}(x, n = 6)
}
\arguments{
\item{x}{an object of the class \code{proteinSeq}}

\item{n}{number of aminoacids to be shown}
}
\description{
The \code{proteinSeq} class is an input for the functions \code{countSeqDiffs} and \code{is.ProteinSeq}. It consists of a character vector. Each entry in this vector represents the aminoacid (the protein components coded by a gene) sequence, for a given aligned protein sequence. The object must be a character, named vector, with the names typically corresponding to the species (name could be scientific or common name) from which every sequence came. The characters within the vector must correspond to valid aminoacid symbols (i.e. capitalized letters or deletion "_" symbols).
}
\examples{

data("cytOxidase")
is.ProteinSeq(cytOxidase)

summary.ProteinSeq(cytOxidase)

head.ProteinSeq(cytOxidase)

tail.ProteinSeq(cytOxidase)

}
\author{
Daniel Rabosky, Matheus Januario, Jennifer Auler
}
