% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_spp}
\alias{mammals_spp}
\title{Mammals species list}
\format{
\subsection{\code{mammals_spp}}{

A \code{vector} with 4099 mammal species.
}
}
\source{
<??>
}
\usage{
mammals_spp
}
\description{
description
}
\keyword{datasets}
