% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_spp}
\alias{mammals_spp}
\title{Mammals species list}
\format{
A \code{vector} containing the names of almost all (i.e., 4099) extant mammal species, following Upham et al (2019) taxonomy. This dataset is part of the package and is licensed
under the Creative Commons Attribution 4.0 International License (CC BY 4.0).
}
\source{
Actual file downloaded from \url{https://vertlife.org/data/}
}
\usage{
mammals_spp
}
\description{
Mammals species list
}
\references{
Upham, N. S., Esselstyn, J. A., & Jetz, W. (2019). Inferring the mammal tree: species-level sets of phylogenies for questions in ecology, evolution, and conservation. PLoS biology, 17(12), e3000494.
}
\concept{data}
\keyword{datasets}
