% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{plotNatSel}
\alias{plotNatSel}
\title{Plot NatSelSim output}
\usage{
plotNatSel(
  gen.HW = gen.HW,
  p.t = p.t,
  w.t = w.t,
  t = t,
  W.gntp = c(w11, w12, w22),
  plot.type = "animateall",
  knitr = FALSE
)
}
\arguments{
\item{gen.HW}{Dataframe with A1A1, A1A2 and A2A2 genotypic
frequencies in each generation (nrows = NGen)}

\item{p.t}{Allelic frequency through time}

\item{w.t}{Mean population fitness through time}

\item{t}{time}

\item{W.gntp}{Initial genotypic fitness}

\item{plot.type}{String indicating if plot should be animated. The default, "animateall", animate all possible panels. Other options are "static", "animate1", "animate3", or "animate4".}

\item{knitr}{Logical indicating if plot is intended to show up in RMarkdown files made by the \code{Knitr} R package.}
}
\value{
Plot of NatSelSim's output (see \code{NatSelSim}'s help page for
details).
}
\description{
Plot NatSelSim output
}
