% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries_fossil}
\alias{timeseries_fossil}
\title{Fossil timeseries}
\format{
\subsection{\code{timeseries_fossil}}{

A \code{data.frame} with 6 colums and 598 rows.
\describe{
\item{clade}{clade}
\item{source}{source}
\item{stem_age}{}
\item{rel_time}{}
\item{time_ma}{time in million year}
\item{richness}{species richness}
}
}
}
\source{
<??>
}
\usage{
timeseries_fossil
}
\description{
description
}
\keyword{datasets}
