% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trilob_fossil}
\alias{trilob_fossil}
\title{Occurrence of trilobites fossils}
\format{
\subsection{\code{trilob_fossil}}{

A \code{data.frame} with 13 colums and 24965 rows.
\describe{
\item{phylum}{phylum}
\item{class}{class}
\item{order}{order}
\item{family}{family}
\item{genus}{genus}
\item{species}{species}
\item{early_interval}{earlier known geological period of ocurence}
\item{late_interval}{later known geological period of ocurence}
\item{max_ma}{maximum known ocurence in million years}
\item{min_ma}{minimun known ocurence in million years}
\item{midpoint}{midpoint}
\item{lng}{longitude}
\item{lat}{latitude}
}
}
}
\source{
\if{html}{\out{
<??>
}}
}
\usage{
trilob_fossil
}
\description{
description
}
\keyword{datasets}
