% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFDriftSim.R
\name{WFDriftSim}
\alias{WFDriftSim}
\title{Simulating generations of genetic drift in a Wright–Fisher (WF) population}
\usage{
WFDriftSim(Ne, nGens, p0 = 0.5, nSim = 1, plot = TRUE, printData = FALSE)
}
\arguments{
\item{Ne}{Number giving the effective population size of the population}

\item{nGens}{Number of generations to be simulated.}

\item{p0}{Initial frequency of a given allele. As the simulated organism is
diploid, the other alleles frequency will be \code{1-(p0)}. Default value
is \code{0.5}.}

\item{nSim}{Number of simulations to be made. If decimals are inserted,
they will be rounded. Default value is \code{1}.}

\item{plot}{Logical indicating if simulations should be plotted as colored
lines. Each color represents a different population. Default value
is \code{TRUE}.}

\item{printData}{Logical indicating whether all simulation results should be
returned as a \code{data.frame}. Default value is \code{FALSE}.}
}
\value{
If \code{plot = TRUE}, plots the timeseries of all simulations,
with each line+color referring to a different simulation. Note that if
many simulations (generally more than 20) are simulated, colors might be
cycled and different simulation will have the same color. If
\code{printData = TRUE}, returns a \code{data.frame} with the
simulation results.
}
\description{
\code{WFDriftSim} simulates genetic drift of diploid Wright–Fisher
populations with a given effective population size through a certain number of
generations.
}
\details{
The effective population size (\code{Ne}) is strongly connected
with the rate of genetic drift (for details, see Waples, 2022)
}
\examples{
#Default values:
WFDriftSim(Ne = 5, nGens = 30)

#A population which has already fixed one of the alleles:
WFDriftSim(Ne = 5, nGens = 30, p0=1)

#Many populations::
WFDriftSim(Ne = 5, nGens = 30, p0=0.2, nSim=20)

######## continuing a previous simulation:
ngen_1stsim <- 10 # number of gens in the 1st sim:
sim1 <- WFDriftSim(Ne = 5, nGens = ngen_1stsim, p0=.2, nSim=20, 
plot = FALSE, printData = TRUE)
ngen_2ndsim <-15 # number of gens in the 2nd sim:
# now, note how we assigned p0:
sim2 <- WFDriftSim(Ne = 5, nGens = ngen_2ndsim, p0=sim1[,ncol(sim1)], 
plot = TRUE, nSim=20, printData = TRUE)

# if we want to merge both simulations, then we have to:
# remove first column of 2nd sim (because it repeats
# the last column of the 1st sim)
sim2 <- sim2[,-1]

# re-name 2nd sim columns:
colnames(sim2) <- paste0("gen", (ngen_1stsim+1):(ngen_1stsim+ngen_2ndsim))

#finally, merging both rounds of simulations:
all_sims <- cbind(sim1, sim2)
head(all_sims)

}
\references{
Fisher RA (1922) On the dominance ratio. Proc. R. Soc. Edinb 42:321–341

Kimura M (1955) Solution of a process of random genetic drift with a
continuous model. PNAS–USA 41(3):144–150

Tran, T. D., Hofrichter, J., & Jost, J. (2013). An introduction to the
mathematical structure of the Wright–Fisher model of population genetics.
Theory in Biosciences, 132(2), 73-82. [good for the historical review,
math can be challenging]

Waples, R. S. (2022). What is Ne, anyway?. Journal of Heredity.

Wright S (1931) Evolution in Mendelian populations. Genetics 16:97–159
}
\author{
Matheus Januario, Dan Rabosky, Jennifer Auler
}
